Fix role index entries to faithfully reproduce the role name
Also prefix the index entries with 'role ' to put all of the roles together
in the index. It was confusing have the base role name scattered around the index
without any idea of what the index entry referred too. This also fixed
some duplicate indexing as LaTeXML considers 'xxx\_yyy' to be the same
as 'xxx yyy'.